refactor(modules): Modules imports es6 style
