+++ src/brainfuck/index.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const _ = require('lodash');
+import _ from 'lodash';
 const overflowPoint = 256;
 
 class Brainfuck {
